Reimplement new logic as extending JDBC lookups, implement basic Consumer logic to update cache.
