+++ druid-lookups/src/main/java/com/yahoo/maha/maha_druid_lookups/server/lookup/namespace/entity/RowMapper.java
@@ -4,7 +4,7 @@ package com.yahoo.maha.maha_druid_lookups.server.lookup.namespace.entity;
 
 import com.metamx.common.logger.Logger;
 import com.yahoo.maha.maha_druid_lookups.query.lookup.namespace.JDBCExtractionNamespace;
-import com.yahoo.maha.maha_druid_lookups.query.lookup.namespace.JDBCProducerExtractionNamespace;
+import com.yahoo.maha.maha_druid_lookups.query.lookup.namespace.JDBCExtractionNamespaceWithLeaderAndFollower;
 import com.yahoo.maha.maha_druid_lookups.query.lookup.namespace.OnlineDatastoreExtractionNamespace;
 import org.skife.jdbi.v2.StatementContext;
 import org.skife.jdbi.v2.tweak.ResultSetMapper;
@@ -22,11 +22,6 @@ public class RowMapper implements ResultSetMapper<Void> {
     private Map<String, List<String>> cache;
 
     public RowMapper(JDBCExtractionNamespace extractionNamespace, Map<String, List<String>> cache) {
-        this.extractionNamespace = extractionNamespace;
-        this.cache = cache;
-    }
-
-    public RowMapper(JDBCProducerExtractionNamespace extractionNamespace, Map<String, List<String>> cache) {
         this.extractionNamespace = extractionNamespace;
         this.cache = cache;
     }
